Create Chapter4CharacterGrid.py
Struggled with this one and eventually had to search the internet/StackOverFlow for help. The instructions stated that x vaules increase to the right and y values increase going down. This seems backwards as the [x][y] format in the instructions shows that the x value actually increases going down and not going to the right. To print the first "O" in the first column, the coordinates are [2][0]. This is an increase from the [0][0] origin point. So, needless to say, this problem was very confusing.